python/requirements/test.txt
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dotenv==1.0.1
PyJWT==2.10.0
httpx>=0.27.0
crypticorn>=2.17.0
